Bug Fix: ValueError in PX4MessageDict.concat (#42)
The original implementation use `TimedeltaIndex` as the index of DataFrame.
However, the default index name of `TimedeltaIndex` is `"timestamp"`, which is duplicate with the column index. Using an index that exist in both index and column will raise a `ValueError`, as `pandas` cannot infer which axis is the user refer to.

In this patch I copy the original DataFrame, set the source index to `None` before calling the `sort_values` and `merge_asof` function. This would resolve the duplication problem.